requirements.txt
fastapi
uvicorn
opencv-python
numpy
Pillow